# apps/api/docs/openapi.yaml
openapi: 3.1.0
info:
  title: AXIOM API
  version: 0.8.0
  description: |
    AXIOM REST API.

    Phase 7 covers the resume lifecycle:
      1. Upload (`POST /api/resumes`) — stores PDF/DOCX in S3/MinIO and kicks off background parsing.
      2. List / read (`GET /api/resumes`, `GET /api/resumes/{id}`) — returns resumes with short-lived presigned URLs.
      3. Delete (`DELETE /api/resumes/{id}`) — removes the resume and its S3 object.
      4. ATS analyze (`POST /api/resumes/{id}/analyze`) — scores a parsed resume against a job description.

    Phase 8 adds the job search + scraping pipeline:
      1. Search (`GET /api/jobs`) — paginated keyword + filter search backed by Postgres ILIKE.
      2. Detail (`GET /api/jobs/{id}`) — single job by ID.
      3. Save / unsave (`POST/DELETE /api/jobs/{id}/save`, `GET /api/jobs/saved`).
      4. Scrape (`POST /api/jobs/scrape`) — admin-only orchestration trigger for Internshala / Unstop / Naukri scrapers (calls AI service which owns the adapter framework).
servers:
  - url: http://localhost:4000
    description: Local development
security:
  - bearerAuth: []
...
paths:
  /api/resumes:
    post:
      tags: [Resumes]
      summary: Upload a resume (PDF or DOCX, max 5 MB)
      operationId: uploadResume
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [resume]
              properties:
                resume:
                  type: string
                  format: binary
      responses:
        "201":
          description: Resume created; parsing runs in the background.
          content:
            application/json:
              schema:
                type: object
                properties:
                  resume: { $ref: "#/components/schemas/Resume" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "413":
          description: File too large (>5 MB).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "415":
          description: Unsupported media type (only PDF / DOCX accepted).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    get:
      tags: [Resumes]
      summary: List the authenticated user's resumes
      operationId: listResumes
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  resumes:
                    type: array
                    items: { $ref: "#/components/schemas/Resume" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /api/jobs:
    get:
      tags: [Jobs]
      summary: Search jobs with filters + pagination
      operationId: searchJobs
      parameters:
        - { in: query, name: q,               schema: { type: string,  maxLength: 200 }, description: Free-text search over title/company/description. }
        - { in: query, name: location,        schema: { type: string,  maxLength: 120 } }
        - { in: query, name: remote,          schema: { type: boolean } }
        - { in: query, name: jobType,         schema: { type: string, enum: [FULL_TIME, PART_TIME, CONTRACT, INTERNSHIP, FREELANCE] } }
        - { in: query, name: experienceLevel, schema: { type: string, enum: [ENTRY, MID, SENIOR, LEAD, EXECUTIVE] } }
        - { in: query, name: source,          schema: { type: string, enum: [internshala, unstop, naukri, manual] } }
        - { in: query, name: salaryMin,       schema: { type: integer, minimum: 0 } }
        - { in: query, name: salaryMax,       schema: { type: integer, minimum: 0 } }
        - { in: query, name: skills,          schema: { type: string }, description: "Comma-separated list of skill names." }
        - { in: query, name: page,            schema: { type: integer, minimum: 1, default: 1 } }
        - { in: query, name: pageSize,        schema: { type: integer, minimum: 1, maximum: 50, default: 20 } }
        - { in: query, name: sortBy,          schema: { type: string, enum: [date, match], default: date } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JobSearchResult" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422":
          description: Validation failed.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ValidationError" }
...
  /api/jobs/{id}:
    parameters:
      - $ref: "#/components/parameters/JobId"
    get:
      tags: [Jobs]
      summary: Get a single job by ID
      operationId: getJob
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties: { job: { $ref: "#/components/schemas/Job" } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /api/jobs/scrape:
    post:
      tags: [Jobs]
      summary: Trigger a scrape run (admin only)
      description: |
        Calls the AI service to scrape `source` for `query`, normalizes the
        results, and upserts into the `jobs` table. Idempotent by unique
        `sourceUrl` — re-runs update existing rows.
      operationId: runScrape
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [source]
              properties:
                source:   { type: string, enum: [internshala, unstop, naukri] }
                query:    { type: string, maxLength: 200, default: "" }
                maxPages: { type: integer, minimum: 1, maximum: 10,  default: 2 }
                maxJobs:  { type: integer, minimum: 1, maximum: 500, default: 100 }
      responses:
        "200":
          description: Scrape complete.
          content:
            application/json:
              schema:
                type: object
                properties:
                  summary: { $ref: "#/components/schemas/ScrapeRunSummary" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422":
          description: Validation failed.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ValidationError" }
        "503":
          description: Scraper service unavailable.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/applications:
    post:
      tags: [Applications]
      summary: Track a new job application
      description: Create a tracker entry for a specific job with initial status.
      operationId: createApplication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [jobId]
              properties:
                jobId:  { type: string, example: cmq111j0000001tyi0aaaaaaa1 }
                status: { type: string, enum: [SAVED, APPLIED, OA_RECEIVED, INTERVIEW_SCHEDULED, OFFER_RECEIVED, REJECTED, WITHDRAWN], default: SAVED }
                note:   { type: string, maxLength: 1000, example: "Bookmarked this job" }
      responses:
        "201":
          description: Tracker entry created successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  application: { $ref: "#/components/schemas/Application" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422":
          description: Validation failed.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ValidationError" }

    get:
      tags: [Applications]
      summary: List user applications
      description: Retrieve user applications with status and date filters.
      operationId: listApplications
      parameters:
        - name: status
          in: query
          required: false
          schema: { type: string, enum: [SAVED, APPLIED, OA_RECEIVED, INTERVIEW_SCHEDULED, OFFER_RECEIVED, REJECTED, WITHDRAWN] }
        - name: dateFrom
          in: query
          required: false
          schema: { type: string, example: "2026-01-01" }
        - name: dateTo
          in: query
          required: false
          schema: { type: string, example: "2026-06-01" }
      responses:
        "200":
          description: Applications list returned.
          content:
            application/json:
              schema:
                type: object
                properties:
                  applications:
                    type: array
                    items: { $ref: "#/components/schemas/Application" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422":
          description: Validation failed.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ValidationError" }
...
components:
  parameters:
...
    JobId:
      name: id
      in: path
      required: true
      description: Job CUID.
      schema: { type: string, example: cmqxxx0000abctyi0aaaa1111 }
  responses:
    BadRequest:
      description: Bad request (e.g. missing file).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing or invalid access token.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Authenticated user does not own this resource.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error: { type: string, example: "Resume not found" }
      required: [error]
    ValidationError:
      type: object
      properties:
        error: { type: string, example: "Validation failed" }
        details:
          type: array
          items:
            type: object
            properties:
              field:   { type: string, example: jobDescription }
              message: { type: string, example: "Job description must be at least 20 characters" }
      required: [error]
    Resume:
      type: object
      required: [id, userId, fileName, fileUrl, fileType, version, createdAt, updatedAt]
      properties:
        id:          { type: string, example: cmqmk6ho00002tyi0w4hl2pph }
        userId:      { type: string }
        fileName:    { type: string, example: cv.pdf }
        fileUrl:     { type: string, format: uri }
        fileType:    { type: string, enum: [pdf, docx] }
        version:     { type: integer, example: 1 }
        parsedData:
          oneOf:
            - { $ref: "#/components/schemas/ParsedResume" }
            - { type: "null" }
        atsScore:
          oneOf:
            - { $ref: "#/components/schemas/ATSScore" }
            - { type: "null" }
        downloadUrl: { type: string, format: uri, description: Presigned download URL (short-lived). }
        createdAt:   { type: string, format: date-time }
        updatedAt:   { type: string, format: date-time }
    ParsedResume:
      type: object
      properties:
        skills:
          type: array
          items:
            type: object
            properties:
              name:        { type: string }
              proficiency:
                oneOf:
                  - { type: string, enum: [beginner, intermediate, advanced, expert] }
                  - { type: "null" }
        experience:
          type: array
          items:
            type: object
            properties:
              company:     { type: string }
              title:       { type: string }
              startDate:   { type: string }
              endDate:     { type: string }
              current:     { type: boolean }
              description: { type: string }
              skills:
                type: array
                items: { type: string }
        education:
          type: array
          items:
            type: object
            properties:
              institution: { type: string }
              degree:      { type: string }
              field:       { type: string }
              startYear:   { type: integer }
              endYear:     { oneOf: [{ type: integer }, { type: "null" }] }
              gpa:         { oneOf: [{ type: number }, { type: "null" }] }
        projects:
          type: array
          items: { type: object }
        certifications:
          type: array
          items: { type: string }
        summary:  { oneOf: [{ type: string }, { type: "null" }] }
        email:    { oneOf: [{ type: string }, { type: "null" }] }
        phone:    { oneOf: [{ type: string }, { type: "null" }] }
        location: { oneOf: [{ type: string }, { type: "null" }] }
    ATSScore:
      type: object
      required: [overall, keywordMatch, completeness, readability, formatting, strengths, missingSkills, suggestions]
      properties:
        overall:       { type: integer, minimum: 0, maximum: 100, example: 79 }
        keywordMatch:  { type: integer, minimum: 0, maximum: 100, example: 76 }
        completeness:  { type: integer, minimum: 0, maximum: 100, example: 100 }
        readability:   { type: integer, minimum: 0, maximum: 100, example: 90 }
        formatting:    { type: integer, minimum: 0, maximum: 100, example: 100 }
        strengths:
          type: array
          items: { type: string }
          example: ["Strong keyword alignment (13 matched terms)"]
        missingSkills:
          type: array
          items: { type: string }
          example: ["kafka", "rabbitmq"]
        suggestions:
          type: array
          items: { type: string }
          example: ["Add a projects section to showcase practical experience"]

    # ── Phase 8: Jobs ────────────────────────────────────────────────────────
    Job:
      type: object
      required: [id, title, company, location, remote, jobType, experienceLevel, source, sourceUrl, postedAt, createdAt, updatedAt]
      properties:
        id:               { type: string,  example: cmq111j0000001tyi0aaaaaaa1 }
        title:            { type: string,  example: "Senior Backend Engineer" }
        company:          { type: string,  example: "Bluefox Systems" }
        companyLogoUrl:   { oneOf: [{ type: string, format: uri }, { type: "null" }] }
        location:         { type: string,  example: "Bengaluru, Karnataka" }
        remote:           { type: boolean, example: false }
        jobType:          { type: string, enum: [FULL_TIME, PART_TIME, CONTRACT, INTERNSHIP, FREELANCE] }
        experienceLevel:  { type: string, enum: [ENTRY, MID, SENIOR, LEAD, EXECUTIVE] }
        salaryMin:        { oneOf: [{ type: integer, minimum: 0 }, { type: "null" }] }
        salaryMax:        { oneOf: [{ type: integer, minimum: 0 }, { type: "null" }] }
        currency:         { type: string, example: INR }
        description:      { type: string }
        requiredSkills:
          type: array
          items: { type: string }
          example: [python, fastapi, postgresql]
        niceToHaveSkills:
          type: array
          items: { type: string }
          example: [kafka, redis]
        source:           { type: string, enum: [internshala, unstop, naukri, manual] }
        sourceUrl:        { type: string, format: uri }
        postedAt:         { type: string, format: date-time }
        expiresAt:        { oneOf: [{ type: string, format: date-time }, { type: "null" }] }
        createdAt:        { type: string, format: date-time }
        updatedAt:        { type: string, format: date-time }
    JobSearchResult:
      type: object
      required: [jobs, total, page, pageSize]
      properties:
        jobs:
          type: array
          items: { $ref: "#/components/schemas/Job" }
        total:    { type: integer, minimum: 0, example: 247 }
        page:     { type: integer, minimum: 1, example: 1 }
        pageSize: { type: integer, minimum: 1, example: 20 }
    ScrapeRunSummary:
      type: object
      required: [source, fetched, inserted, updated, skipped, errors, durationMs]
      properties:
        source:     { type: string, enum: [internshala, unstop, naukri] }
        fetched:    { type: integer, example: 18 }
        inserted:   { type: integer, example: 14 }
        updated:    { type: integer, example: 4 }
        skipped:    { type: integer, example: 0 }
        errors:     { type: integer, example: 0 }
        durationMs: { type: integer, example: 4321 }
    Application:
      type: object
      required: [id, userId, jobId, status, timeline, createdAt, updatedAt]
      properties:
        id:          { type: string, example: cmaaaa0000000000000000000 }
        userId:      { type: string }
        jobId:       { type: string }
        status:      { type: string, enum: [SAVED, APPLIED, OA_RECEIVED, INTERVIEW_SCHEDULED, OFFER_RECEIVED, REJECTED, WITHDRAWN] }
        coverLetter: { type: string }
        notes:       { type: string }
        appliedAt:   { oneOf: [{ type: string, format: date-time }, { type: "null" }] }
        timeline:
          type: array
          items: { $ref: "#/components/schemas/ApplicationTimelineEntry" }
        job:         { $ref: "#/components/schemas/Job" }
        createdAt:   { type: string, format: date-time }
        updatedAt:   { type: string, format: date-time }
    ApplicationTimelineEntry:
      type: object
      required: [status, at, note]
      properties:
        status: { type: string, enum: [SAVED, APPLIED, OA_RECEIVED, INTERVIEW_SCHEDULED, OFFER_RECEIVED, REJECTED, WITHDRAWN] }
        at:     { type: string, format: date-time }
        note:   { type: string }
...